fix: postgre data from float to decimal

+++ migrations/001_create_users_table.up.sql
@@ -1,4 +1,4 @@
 CREATE TABLE users (
     id SERIAL PRIMARY KEY,
-    balance FLOAT
+    balance DECIMAL
 );
